137: Add graceful shutdown, update .gitignores
